+++ api-number-at-letters/changelog.md
@@ -11,3 +11,4 @@
 
 ## [0.1.1]
 - Refactor response objects and enable global filter exceptions
+- Enable option bearer token in swagger ui and service description